test: split gateway org-policy tests to their own file (max-lines gate)
